--- src/components/Signin.js
@@ -1,54 +1,0 @@
-import React, { useState } from "react";
-import axios from "axios";
-import { useNavigate } from "react-router-dom";
-
-const Signin = () => {
-  const [formData, setFormData] = useState({ email: "", password: "" });
-  const [errorMessage, setErrorMessage] = useState("");
-  const navigate = useNavigate();
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await axios.post("http://localhost:3001/users/login", formData);
-      console.log(response);
-      navigate("/quiz"); // Navigate to the homepage or next page after successful login
-    } catch (err) {
-      setErrorMessage(err.response?.data?.error || "Invalid credentials.");
-    }
-  };
-
-  return (
-    <div className="signin-container">
-      <h2>Sign In</h2>
-      <form onSubmit={handleSubmit}>
-        <input
-          type="email"
-          name="email"
-          value={formData.email}
-          onChange={handleChange}
-          placeholder="Email"
-          required
-        />
-        <input
-          type="password"
-          name="password"
-          value={formData.password}
-          onChange={handleChange}
-          placeholder="Password"
-          required
-        />
-        <button type="submit">Sign In</button>
-      </form>
-      {errorMessage && <p className="error-message">{errorMessage}</p>}
-      <button onClick={() => navigate("/forgot-password")}>Forgot Password?</button>
-    </div>
-  );
-};
-
-export default Signin;
